Initial Provisioning Script
Modified the base ansible script that I previously used for my personal website, to now be able to quickly provision (sort of, the installation and updating of the server takes a while) the server that will run the HormoziGPT project. The playbook installs dependencies, Pyenv, and then the appropriate python version, and finally clones and configures the HormoziGPT project. Now its just a matter of remoting in and trying it out.

Also created a very simple Readme file to detail the different steps to set up the server.