Reverse Proxy and SSL
Set up the configuration for the reverse proxy for the demo. The configuration is loaded by the provisioning script and its then applied to create a route into the server. However, the connection would not be considered "secure" if you just leave it like that, so I also set up certbot, to add SSL certificates to the server. Now its actually working. You just need to go into the server, create a tmux window and start the streamlit application. Its still not properly loading, so thats the next thing I will troubleshoot.