Notebook for Adding Data to Pinecone
In the video used as reference: https://www.youtube.com/watch?v=86Gw317oEHk&t=618s, the dude doesn't provide the notebook to actually populate the pinecone database for his application. Here I implemented a quick notebook that takes a series of text files, properly chops them up into chunks, adds embeddings with the OpenAI API, and finally stores the data in Pinecone. It was relatively simple, but I havent tested it. I include the 3 sample files that I used to generate 9 vectors. Also its worth mentioning that I added poetry to handle all of the dependencies for the project. Its simpler and easier to manage.